requirements.txt
groq==0.9.0
python-dotenv==1.0.0
deepgram-sdk==3.5.0
sounddevice==0.4.6
numpy==1.24.3
requests==2.31.0
pytest==7.4.0
asyncio-contextmanager==1.0.1
websockets==11.0.3
aenum==3.1.16
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comtypes==1.4.15
dataclasses-json==0.6.7
deepgram-sdk==3.5.0
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.3.1
frozenlist==1.8.0
groq==0.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Js2Py==0.74
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.24.3
packaging==26.0
pip==26.0.1
pipwin==0.5.2
pluggy==1.6.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyjsparser==2.7.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pytest==7.4.0
python-dotenv==1.0.0
pyttsx3==2.99
pywin32==311
requests==2.31.0
setuptools==57.4.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.8.3
tomli==2.4.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0
yarl==1.22.0
